+++ tests/superuser-secret-generated/superuser_secret_generated_test.go
@@ -9,7 +9,7 @@ import (
 
 	"encoding/base64"
 
-	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
 	"github.com/k8ssandra/cass-operator/tests/kustomize"