+++ tests/terminate/terminate_suite_test.go
@@ -9,7 +9,7 @@ import (
 	"sync"
 	"testing"
 
-	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
 	"github.com/k8ssandra/cass-operator/tests/kustomize"