+++ tests/timeout_prestop_termination/timeout_prestop_termination_suite_test.go
@@ -11,7 +11,7 @@ import (
 	"testing"
 	"time"
 
-	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
 	"github.com/k8ssandra/cass-operator/tests/kustomize"