+++ scripts/pre-release-process.sh
@@ -26,3 +26,12 @@ cd config/manager && $KUSTOMIZE edit set image controller=$IMG && cd -
 
 # Modify config/manager/image_config.yaml to have proper version for server-system-logger
 LOG_IMG=k8ssandra/system-logger:${TAG} yq eval -i '.images.system-logger = env(LOG_IMG)' config/manager/image_config.yaml
+
+# Now add everything and create a commit + tag
+git add CHANGELOG.md
+git add README.md
+git add config/manager/kustomization.yaml
+git add config/manager/image_config.yaml
+
+git commit -m "Release $TAG"
+git tag -a $TAG -m "Release $TAG"