+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Changelog for Cass Operator, new PRs should update the `main / unreleased` secti
 
 ## unreleased
 
+* [CHANGE] [#291](https://github.com/k8ssandra/cass-operator/issues/291) Update Ginkgo to v2 (maintain current features, nothing additional from v2)
+
 ## v1.13.0
 
 * [CHANGE] [#395](https://github.com/k8ssandra/cass-operator/issues/395) Make CassandraTask job arguments strongly typed